+++ evals/tpcds_agent/query_prompts.json
@@ -136,7 +136,7 @@
   {
     "id": 26,
     "grade": "easy",
-    "prompt": "For catalog sales in the year 2000, report the average quantity, average list price, average coupon amount, and average sales price for each item. Filter by the customer-demographic profile recorded on each catalog sale itself (the one linked directly to the catalog sale's billing customer-demographic key, NOT the customer's current demographic profile reached through the customer chain): male, single, 'College' education. Also restrict to promotions where the email channel flag is 'N' or the event channel flag is 'N' (at this scale factor these flags are always 'N', so the clause matches every promotion). Order by item code (the item's business identifier); limit 100 rows."
+    "prompt": "For catalog sales in the year 2000, report the average quantity, average list price, average coupon amount, and average sales price for each item. Filter by the point-of-sale customer-demographic profile recorded directly on each catalog sale: male, single, 'College' education. Do not use the billing customer's current demographic profile. Also restrict to promotions where the email channel flag is 'N' or the event channel flag is 'N' (at this scale factor these flags are always 'N', so the clause matches every promotion). Order by item code (the item's business identifier); limit 100 rows."
   },
   {
     "id": 27,
