+++ tests/modeling/tpc_ds_duckdb/query26.preql
@@ -4,9 +4,9 @@
 import catalog_sales as cs;
 
 where
-    cs.billing_customer_demographic.gender = 'M'
-    and cs.billing_customer_demographic.marital_status = 'S'
-    and cs.billing_customer_demographic.education_status = 'College'
+    cs.pos_customer_demographic.gender = 'M'
+    and cs.pos_customer_demographic.marital_status = 'S'
+    and cs.pos_customer_demographic.education_status = 'College'
     and (cs.promotion.channel_email = 'N' or cs.promotion.channel_event = 'N')
     and cs.date.year = 2000
 select
