+++ tests/modeling/tpc_ds_duckdb/zquery26.log
@@ -1,7 +1,7 @@
 query_id = "26"
 generator = "v3"
-gen_length = 1405
-preql_size = 509
+gen_length = 1385
+preql_size = 497
 comp_size = 583
 generated_sql = """
 SELECT
@@ -15,9 +15,9 @@ FROM
     INNER JOIN \"memory\".\"date_dim\" as \"cs_date_date\" on \"cs_catalog_sales\".\"CS_SOLD_DATE_SK\" = \"cs_date_date\".\"D_DATE_SK\"
     INNER JOIN \"memory\".\"item\" as \"cs_item_items\" on \"cs_catalog_sales\".\"CS_ITEM_SK\" = \"cs_item_items\".\"I_ITEM_SK\"
     INNER JOIN \"memory\".\"promotion\" as \"cs_promotion_promotion\" on \"cs_catalog_sales\".\"CS_PROMO_SK\" = \"cs_promotion_promotion\".\"P_PROMO_SK\"
-    INNER JOIN \"memory\".\"customer_demographics\" as \"cs_billing_customer_demographic_customer_demographics\" on \"cs_catalog_sales\".\"CS_BILL_CDEMO_SK\" = \"cs_billing_customer_demographic_customer_demographics\".\"CD_DEMO_SK\"
+    INNER JOIN \"memory\".\"customer_demographics\" as \"cs_pos_customer_demographic_customer_demographics\" on \"cs_catalog_sales\".\"CS_BILL_CDEMO_SK\" = \"cs_pos_customer_demographic_customer_demographics\".\"CD_DEMO_SK\"
 WHERE
-    \"cs_billing_customer_demographic_customer_demographics\".\"CD_GENDER\" = 'M' and \"cs_billing_customer_demographic_customer_demographics\".\"CD_MARITAL_STATUS\" = 'S' and \"cs_billing_customer_demographic_customer_demographics\".\"CD_EDUCATION_STATUS\" = 'College' and ( \"cs_promotion_promotion\".\"P_CHANNEL_EMAIL\" = 'N' or \"cs_promotion_promotion\".\"P_CHANNEL_EVENT\" = 'N' ) and \"cs_date_date\".\"D_YEAR\" = 2000
+    \"cs_pos_customer_demographic_customer_demographics\".\"CD_GENDER\" = 'M' and \"cs_pos_customer_demographic_customer_demographics\".\"CD_MARITAL_STATUS\" = 'S' and \"cs_pos_customer_demographic_customer_demographics\".\"CD_EDUCATION_STATUS\" = 'College' and ( \"cs_promotion_promotion\".\"P_CHANNEL_EMAIL\" = 'N' or \"cs_promotion_promotion\".\"P_CHANNEL_EVENT\" = 'N' ) and \"cs_date_date\".\"D_YEAR\" = 2000
 
 GROUP BY
     1
