+++ tests/modeling/tpc_ds_duckdb/zquery_timing_amd64-intel64_family_6_model_183_stepping_1_genuineintel-16.log
@@ -84,9 +84,9 @@ exec_time = 0.12700349999977334
 comp_time = 0.10947189999978946
 
 [query_18]
-parse_time = 0.3234010999999555
-exec_time = 0.09692930000028355
-comp_time = 0.08821769999985918
+parse_time = 0.16681320000043343
+exec_time = 0.10309800000050018
+comp_time = 0.22842169999967155
 
 [query_19]
 parse_time = 0.06939079999983733
@@ -134,9 +134,9 @@ exec_time = 0.10991749999993772
 comp_time = 0.04439349999984188
 
 [query_26]
-parse_time = 0.06950340000003052
-exec_time = 0.06654549999984738
-comp_time = 0.05555560000038895
+parse_time = 0.07346310000048106
+exec_time = 0.06831199999942328
+comp_time = 0.055570699999407225
 
 [query_27]
 parse_time = 0.0552626999997301
@@ -369,9 +369,9 @@ exec_time = 0.036644300000261865
 comp_time = 0.03432969999994384
 
 [query_72]
-parse_time = 0.26413370000000214
-exec_time = 0.044900500000039756
-comp_time = 22.401872599999933
+parse_time = 0.2564710999995441
+exec_time = 0.03244360000007873
+comp_time = 24.26367669999945
 
 [query_73]
 parse_time = 0.18151339999985794
